New license chooser API
The nautilus extension still uses it's own code to select a license, although an example of how it would be used in python is available in bindings/python/test.py.  The KDE4 license chooser does use this new API.

I'm going to go ahead and commit this without discussion because it's a non-invasive change, although I'd like some feedback.


git-svn-id: https://cctools.svn.sourceforge.net/svnroot/cctools/liblicense/trunk@6428 aec4f576-4910-0410-beeb-ecb0427469e7